requirements.txt
flask
opencv-python
tesseract
paddlepaddle
paddleocr
requests
textract